fix: build errors after the refactor

+++ src/robotfindskitten/engine/details/base_types.hpp
@@ -40,7 +40,7 @@ class sprite {
     // Expect CharT null terminator
     std::basic_string<char_size> data_;
     [[nodiscard]] static sprite normilize_from_string(std::string s);
-    sprite(uint16_t width, std::basic_string<char_size> str)
+    sprite(size_t width, std::basic_string<char_size> str)
         : width_(width), data_(std::move(str)) {}
 
 public:
